Add code splitting for faster load time and fix imports
