+++ .gitignore
@@ -2,6 +2,7 @@
 node_modules/
 public_wallet_app/js/cardano.bundle.js
 public_wallet_app/js/frontend.bundle.js
+public_wallet_app/js/*.frontend.bundle.js
 test/browser/js/cardano.bundle.test.js
 .env
 yarn-error.log
