+++ server/helpers/loadFrontendConfig.js
@@ -6,4 +6,5 @@ module.exports = {
   CARDANOLITE_TRANSACTION_SUBMITTER_URL: process.env.CARDANOLITE_TRANSACTION_SUBMITTER_URL,
   CARDANOLITE_WALLET_ADDRESS_LIMIT: parseInt(process.env.CARDANOLITE_WALLET_ADDRESS_LIMIT, 10),
   CARDANOLITE_DEMO_WALLET_MNEMONIC: process.env.CARDANOLITE_DEMO_WALLET_MNEMONIC,
+  CARDANOLITE_ENABLE_DEBUGGING: process.env.CARDANOLITE_ENABLE_DEBUGGING,
 }
